Reorganize tests to co-location with source files
Move test files from tests/ directory to be co-located with source files
and split types-related tests into a separate file:

- tests/test_utils.py → kubeflow/trainer/utils/utils_test.py
- Extract types tests → kubeflow/trainer/types/types_test.py
- Update pyproject.toml testpaths: ["tests"] → ["kubeflow"]
- Remove tests/ directory

This improves code organization by keeping tests next to the code
they validate, making it easier to maintain test coverage when
modifying source files.
