+++ python/pyproject.toml
@@ -61,7 +61,7 @@ addopts = [
     "-v",
     "--tb=short",
 ]
-testpaths = ["tests"]
+testpaths = ["kubeflow"]
 python_files = ["test_*.py", "*_test.py"]
 python_classes = ["Test*"]
 python_functions = ["test_*"]
